APIC MSRs setup
