+++ bfvmm/include/intrinsics/x86/common/msrs_x64.h
@@ -98,151 +98,6 @@ namespace ia32_tsc
     { bfdebug_nhex(level, name, get(), msg); }
 }
 
-namespace ia32_apic_base
-{
-    constexpr const auto addr = 0x0000001BU;
-    constexpr const auto name = "ia32_apic_base";
-
-    inline auto get() noexcept
-    { return _read_msr(addr); }
-
-    inline void set(value_type val) noexcept
-    { _write_msr(addr, val); }
-
-    namespace bsp_flag
-    {
-        constexpr const auto mask = 0x0000000000000100ULL;
-        constexpr const auto from = 8ULL;
-        constexpr const auto name = "bsp_flag";
-
-        inline auto is_enabled()
-        { return is_bit_set(_read_msr(addr), from); }
-
-        inline auto is_enabled(value_type msr)
-        { return is_bit_set(msr, from); }
-
-        inline auto is_disabled()
-        { return is_bit_cleared(_read_msr(addr), from); }
-
-        inline auto is_disabled(value_type msr)
-        { return is_bit_cleared(msr, from); }
-
-        inline void enable()
-        { _write_msr(addr, set_bit(_read_msr(addr), from)); }
-
-        inline auto enable(value_type msr)
-        { return set_bit(msr, from); }
-
-        inline void disable()
-        { _write_msr(addr, clear_bit(_read_msr(addr), from)); }
-
-        inline auto disable(value_type msr)
-        { return clear_bit(msr, from); }
-
-        inline void dump(int level, std::string *msg = nullptr)
-        { bfdebug_subbool(level, name, is_enabled(), msg); }
-    }
-
-    namespace enable_x2apic
-    {
-        constexpr const auto mask = 0x0000000000000400ULL;
-        constexpr const auto from = 10ULL;
-        constexpr const auto name = "enable_x2apic";
-
-        inline auto is_enabled()
-        { return is_bit_set(_read_msr(addr), from); }
-
-        inline auto is_enabled(value_type msr)
-        { return is_bit_set(msr, from); }
-
-        inline auto is_disabled()
-        { return is_bit_cleared(_read_msr(addr), from); }
-
-        inline auto is_disabled(value_type msr)
-        { return is_bit_cleared(msr, from); }
-
-        inline void enable()
-        { _write_msr(addr, set_bit(_read_msr(addr), from)); }
-
-        inline auto enable(value_type msr)
-        { return set_bit(msr, from); }
-
-        inline void disable()
-        { _write_msr(addr, clear_bit(_read_msr(addr), from)); }
-
-        inline auto disable(value_type msr)
-        { return clear_bit(msr, from); }
-
-        inline void dump(int level, std::string *msg = nullptr)
-        { bfdebug_subbool(level, name, is_enabled(), msg); }
-    }
-
-    namespace apic_global_enable
-    {
-        constexpr const auto mask = 0x0000000000000800ULL;
-        constexpr const auto from = 11ULL;
-        constexpr const auto name = "apic_global_enable";
-
-        inline auto is_enabled()
-        { return is_bit_set(_read_msr(addr), from); }
-
-        inline auto is_enabled(value_type msr)
-        { return is_bit_set(msr, from); }
-
-        inline auto is_disabled()
-        { return is_bit_cleared(_read_msr(addr), from); }
-
-        inline auto is_disabled(value_type msr)
-        { return is_bit_cleared(msr, from); }
-
-        inline void enable()
-        { _write_msr(addr, set_bit(_read_msr(addr), from)); }
-
-        inline auto enable(value_type msr)
-        { return set_bit(msr, from); }
-
-        inline void disable()
-        { _write_msr(addr, clear_bit(_read_msr(addr), from)); }
-
-        inline auto disable(value_type msr)
-        { return clear_bit(msr, from); }
-
-        inline void dump(int level, std::string *msg = nullptr)
-        { bfdebug_subbool(level, name, is_enabled(), msg); }
-    }
-
-    namespace apic_base
-    {
-        constexpr const auto mask = 0xFFFFFFFFFFFFF000ULL;
-        constexpr const auto from = 12ULL;
-        constexpr const auto name = "apic_base";
-
-        inline auto get() noexcept
-        { return get_bits(_read_msr(addr), mask) >> from; }
-
-        inline auto get(value_type msr) noexcept
-        { return get_bits(msr, mask) >> from; }
-
-        inline void set(value_type val) noexcept
-        { _write_msr(addr, set_bits(_read_msr(addr), mask, val << from)); }
-
-        inline auto set(value_type msr, value_type val) noexcept
-        { return set_bits(msr, mask, val << from); }
-
-        inline void dump(int level, std::string *msg = nullptr)
-        { bfdebug_subnhex(level, name, get(), msg); }
-    }
-
-    inline void dump(int level, std::string *msg = nullptr)
-    {
-        bfdebug_nhex(level, name, get(), msg);
-        bsp_flag::dump(level, msg);
-        enable_x2apic::dump(level, msg);
-        apic_global_enable::dump(level, msg);
-        apic_base::dump(level, msg);
-    }
-}
-
 namespace ia32_mperf
 {
     constexpr const auto addr = 0x000000E7U;
