+++ bfvmm/src/intrinsics/tests/CMakeLists.txt
@@ -31,6 +31,8 @@ macro(do_test str)
     add_test(test_${str} test_${str})
 endmacro(do_test)
 
+do_test(apic_intel_x64_lapic)
+do_test(apic_intel_x64_x2apic)
 do_test(cache_x64)
 do_test(cpuid_x64)
 do_test(crs_intel_x64)
