+++ bfvmm/src/intrinsics/tests/test_apic_intel_x64_x2apic.cpp
@@ -1,0 +1,1419 @@
+//
+// Bareflank Hypervisor
+// Copyright (C) 2015 Assured Information Security, Inc.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+
+#include <catch/catch.hpp>
+#include <intrinsics/x86/intel_x64.h>
+#include <hippomocks.h>
+
+#ifdef _HIPPOMOCKS__ENABLE_CFUNC_MOCKING_SUPPORT
+
+using namespace x64;
+
+TEST_CASE("test name goes here")
+{
+    CHECK(true);
+}
+
+std::map<msrs::field_type, msrs::value_type> g_msrs;
+
+extern "C" uint64_t
+test_read_msr(uint32_t addr) noexcept
+{ return g_msrs[addr]; }
+
+extern "C" void
+test_write_msr(uint32_t addr, uint64_t val) noexcept
+{ g_msrs[addr] = val; }
+
+static void
+setup_intrinsics(MockRepository &mocks)
+{
+    mocks.OnCallFunc(_read_msr).Do(test_read_msr);
+    mocks.OnCallFunc(_write_msr).Do(test_write_msr);
+}
+
+TEST_CASE("ia32_x2apic_apicid")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_apicid;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_version")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_version;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tpr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tpr;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_ppr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_ppr;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_eoi")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_eoi;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(g_msrs[addr] == 0xFFFFFFFFFFFFFFFFULL);
+}
+
+TEST_CASE("ia32_x2apic_ldr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_ldr;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_ldr_logical_id")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_ldr;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(logical_id::get() == (logical_id::mask >> logical_id::from));
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(logical_id::get(logical_id::mask) == (logical_id::mask >> logical_id::from));
+}
+
+TEST_CASE("ia32_x2apic_ldr_cluster_id")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_ldr;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(cluster_id::get() == (cluster_id::mask >> cluster_id::from));
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(cluster_id::get(cluster_id::mask) == (cluster_id::mask >> cluster_id::from));
+}
+
+TEST_CASE("ia32_x2apic_sivr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_sivr;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_sivr_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_sivr;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_sivr_apic_enable_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_sivr;
+
+    apic_enable_bit::enable();
+    CHECK(apic_enable_bit::is_enabled());
+    apic_enable_bit::disable();
+    CHECK(apic_enable_bit::is_disabled());
+
+    apic_enable_bit::enable(apic_enable_bit::mask);
+    CHECK(apic_enable_bit::is_enabled(apic_enable_bit::mask));
+    apic_enable_bit::disable(0x0);
+    CHECK(apic_enable_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_sivr_focus_checking")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_sivr;
+
+    focus_checking::enable();
+    CHECK(focus_checking::is_enabled());
+    focus_checking::disable();
+    CHECK(focus_checking::is_disabled());
+
+    focus_checking::enable(0x0);
+    CHECK(focus_checking::is_enabled(0x0));
+    focus_checking::disable(focus_checking::mask);
+    CHECK(focus_checking::is_disabled(focus_checking::mask));
+}
+
+TEST_CASE("ia32_x2apic_sivr_suppress_eoi_broadcast")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_sivr;
+
+    suppress_eoi_broadcast::enable();
+    CHECK(suppress_eoi_broadcast::is_enabled());
+    suppress_eoi_broadcast::disable();
+    CHECK(suppress_eoi_broadcast::is_disabled());
+
+    suppress_eoi_broadcast::enable(suppress_eoi_broadcast::mask);
+    CHECK(suppress_eoi_broadcast::is_enabled(suppress_eoi_broadcast::mask));
+    suppress_eoi_broadcast::disable(0x0);
+    CHECK(suppress_eoi_broadcast::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_isr0")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr0;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr1")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr1;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr2")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr2;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr3")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr3;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr4")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr4;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr5")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr5;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr6")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr6;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_isr7")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_isr7;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr0")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr0;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr1")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr1;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr2")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr2;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr3")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr3;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr4")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr4;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr5")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr5;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr6")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr6;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_tmr7")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_tmr7;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr0")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr0;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr1")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr1;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr2")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr2;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr3")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr3;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr4")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr4;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr5")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr5;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr6")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr6;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_irr7")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_irr7;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_esr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_esr;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_cmci")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_cmci;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_cmci_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_cmci;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_cmci_delivery_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_cmci;
+
+    delivery_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get() == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(delivery_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get(delivery_mode::mask) == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(0x0ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x2ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x4ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x5ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x7ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x1ULL);
+    delivery_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_cmci_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_cmci;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_cmci_mask_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_cmci;
+
+    mask_bit::enable();
+    CHECK(mask_bit::is_enabled());
+    mask_bit::disable();
+    CHECK(mask_bit::is_disabled());
+
+    mask_bit::enable(mask_bit::mask);
+    CHECK(mask_bit::is_enabled(mask_bit::mask));
+    mask_bit::disable(0x0);
+    CHECK(mask_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_icr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_icr_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_icr_delivery_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    delivery_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get() == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(delivery_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get(delivery_mode::mask) == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(0x0ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x2ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x4ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x5ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x7ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x1ULL);
+    delivery_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_icr_destination_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    destination_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(destination_mode::get() == (destination_mode::mask >> destination_mode::from));
+
+    destination_mode::set(destination_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(destination_mode::get(destination_mode::mask) == (destination_mode::mask >> destination_mode::from));
+
+    destination_mode::set(0x0000000000000000ULL);
+    dump(0);
+    destination_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_icr_level")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    level::enable();
+    CHECK(level::is_enabled());
+    level::disable();
+    CHECK(level::is_disabled());
+
+    level::enable(level::mask);
+    CHECK(level::is_enabled(level::mask));
+    level::disable(0x0);
+    CHECK(level::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_icr_trigger_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    trigger_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(trigger_mode::get() == (trigger_mode::mask >> trigger_mode::from));
+
+    trigger_mode::set(trigger_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(trigger_mode::get(trigger_mode::mask) == (trigger_mode::mask >> trigger_mode::from));
+
+    trigger_mode::set(0x0000000000000000ULL);
+    dump(0);
+    trigger_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_icr_destination_shorthand")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    destination_shorthand::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(destination_shorthand::get() == (destination_shorthand::mask >> destination_shorthand::from));
+
+    destination_shorthand::set(destination_shorthand::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(destination_shorthand::get(destination_shorthand::mask) == (destination_shorthand::mask >> destination_shorthand::from));
+
+    destination_shorthand::set(0x0ULL);
+    destination_shorthand::dump(0);
+    destination_shorthand::set(0x1ULL);
+    destination_shorthand::dump(0);
+    destination_shorthand::set(0x2ULL);
+    destination_shorthand::dump(0);
+    destination_shorthand::set(0x3ULL);
+    destination_shorthand::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_icr_destination_field")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_icr;
+
+    destination_field::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(destination_field::get() == (destination_field::mask >> destination_field::from));
+
+    destination_field::set(destination_field::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(destination_field::get(destination_field::mask) == (destination_field::mask >> destination_field::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_timer")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_timer;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_timer_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_timer;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_timer_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_timer;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_timer_mask_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_timer;
+
+    mask_bit::enable();
+    CHECK(mask_bit::is_enabled());
+    mask_bit::disable();
+    CHECK(mask_bit::is_disabled());
+
+    mask_bit::enable(mask_bit::mask);
+    CHECK(mask_bit::is_enabled(mask_bit::mask));
+    mask_bit::disable(0x0);
+    CHECK(mask_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_lvt_timer_timer_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_timer;
+
+    timer_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(timer_mode::get() == (timer_mode::mask >> timer_mode::from));
+
+    timer_mode::set(timer_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(timer_mode::get(timer_mode::mask) == (timer_mode::mask >> timer_mode::from));
+
+    timer_mode::set(0x0ULL);
+    timer_mode::dump(0);
+    timer_mode::set(0x1ULL);
+    timer_mode::dump(0);
+    timer_mode::set(0x2ULL);
+    timer_mode::dump(0);
+    timer_mode::set(0x3ULL);
+    timer_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_thermal")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_thermal;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_thermal_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_thermal;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_thermal_delivery_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_thermal;
+
+    delivery_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get() == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(delivery_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get(delivery_mode::mask) == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(0x0ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x2ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x4ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x5ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x7ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x1ULL);
+    delivery_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_thermal_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_thermal;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_thermal_mask_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_thermal;
+
+    mask_bit::enable();
+    CHECK(mask_bit::is_enabled());
+    mask_bit::disable();
+    CHECK(mask_bit::is_disabled());
+
+    mask_bit::enable(mask_bit::mask);
+    CHECK(mask_bit::is_enabled(mask_bit::mask));
+    mask_bit::disable(0x0);
+    CHECK(mask_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_lvt_pmi")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_pmi;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_pmi_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_pmi;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_pmi_delivery_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_pmi;
+
+    delivery_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get() == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(delivery_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get(delivery_mode::mask) == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(0x0ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x2ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x4ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x5ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x7ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x1ULL);
+    delivery_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_pmi_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_pmi;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_pmi_mask_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_pmi;
+
+    mask_bit::enable();
+    CHECK(mask_bit::is_enabled());
+    mask_bit::disable();
+    CHECK(mask_bit::is_disabled());
+
+    mask_bit::enable(mask_bit::mask);
+    CHECK(mask_bit::is_enabled(mask_bit::mask));
+    mask_bit::disable(0x0);
+    CHECK(mask_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_delivery_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    delivery_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get() == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(delivery_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get(delivery_mode::mask) == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(0x0ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x2ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x4ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x5ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x7ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x1ULL);
+    delivery_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_polarity")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    polarity::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(polarity::get() == (polarity::mask >> polarity::from));
+
+    polarity::set(polarity::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(polarity::get(polarity::mask) == (polarity::mask >> polarity::from));
+
+    polarity::set(0x0000000000000000ULL);
+    dump(0);
+    polarity::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_remote_irr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(remote_irr::get() == (remote_irr::mask >> remote_irr::from));
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(remote_irr::get(remote_irr::mask) == (remote_irr::mask >> remote_irr::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_trigger_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    trigger_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(trigger_mode::get() == (trigger_mode::mask >> trigger_mode::from));
+
+    trigger_mode::set(trigger_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(trigger_mode::get(trigger_mode::mask) == (trigger_mode::mask >> trigger_mode::from));
+
+    trigger_mode::set(0x0000000000000000ULL);
+    dump(0);
+    trigger_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint0_mask_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint0;
+
+    mask_bit::enable();
+    CHECK(mask_bit::is_enabled());
+    mask_bit::disable();
+    CHECK(mask_bit::is_disabled());
+
+    mask_bit::enable(mask_bit::mask);
+    CHECK(mask_bit::is_enabled(mask_bit::mask));
+    mask_bit::disable(0x0);
+    CHECK(mask_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_delivery_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    delivery_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get() == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(delivery_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_mode::get(delivery_mode::mask) == (delivery_mode::mask >> delivery_mode::from));
+
+    delivery_mode::set(0x0ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x2ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x4ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x5ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x7ULL);
+    delivery_mode::dump(0);
+    delivery_mode::set(0x1ULL);
+    delivery_mode::dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_polarity")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    polarity::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(polarity::get() == (polarity::mask >> polarity::from));
+
+    polarity::set(polarity::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(polarity::get(polarity::mask) == (polarity::mask >> polarity::from));
+
+    polarity::set(0x0000000000000000ULL);
+    dump(0);
+    polarity::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_remote_irr")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(remote_irr::get() == (remote_irr::mask >> remote_irr::from));
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(remote_irr::get(remote_irr::mask) == (remote_irr::mask >> remote_irr::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_trigger_mode")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    trigger_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(trigger_mode::get() == (trigger_mode::mask >> trigger_mode::from));
+
+    trigger_mode::set(trigger_mode::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(trigger_mode::get(trigger_mode::mask) == (trigger_mode::mask >> trigger_mode::from));
+
+    trigger_mode::set(0x0000000000000000ULL);
+    dump(0);
+    trigger_mode::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_lint1_mask_bit")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_lint1;
+
+    mask_bit::enable();
+    CHECK(mask_bit::is_enabled());
+    mask_bit::disable();
+    CHECK(mask_bit::is_disabled());
+
+    mask_bit::enable(mask_bit::mask);
+    CHECK(mask_bit::is_enabled(mask_bit::mask));
+    mask_bit::disable(0x0);
+    CHECK(mask_bit::is_disabled(0x0));
+}
+
+TEST_CASE("ia32_x2apic_lvt_error")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_error;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_lvt_error_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_error;
+
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get() == (vector::mask >> vector::from));
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(vector::get(vector::mask) == (vector::mask >> vector::from));
+}
+
+TEST_CASE("ia32_x2apic_lvt_error_delivery_status")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_lvt_error;
+
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get() == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(delivery_status::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(delivery_status::get(delivery_status::mask) == (delivery_status::mask >> delivery_status::from));
+
+    delivery_status::set(0x0000000000000000ULL);
+    dump(0);
+    delivery_status::set(0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_init_count")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_init_count;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_cur_count")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_cur_count;
+
+    g_msrs[addr] = 0xFFFFFFFFFFFFFFFFULL;
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_div_conf")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_div_conf;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
+    dump(0);
+}
+
+TEST_CASE("ia32_x2apic_self_ipi")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_self_ipi;
+
+    set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(g_msrs[addr] == 0xFFFFFFFFFFFFFFFFULL);
+}
+
+TEST_CASE("ia32_x2apic_self_ipi_vector")
+{
+    MockRepository mocks;
+    setup_intrinsics(mocks);
+
+    using namespace intel_x64::msrs::ia32_x2apic_self_ipi;
+
+    set(0x0000000000000000ULL);
+    vector::set(0xFFFFFFFFFFFFFFFFULL);
+    CHECK(g_msrs[addr] == vector::mask);
+
+    vector::set(vector::mask, 0xFFFFFFFFFFFFFFFFULL);
+    CHECK(g_msrs[addr] == vector::mask);
+}
+
+#endif
