+++ bfvmm/src/intrinsics/tests/test_msrs_x64.cpp
@@ -82,86 +82,6 @@ TEST_CASE("ia32_tsc")
     dump(0);
 }
 
-TEST_CASE("ia32_apic_base")
-{
-    MockRepository mocks;
-    setup_intrinsics(mocks);
-
-    using namespace msrs::ia32_apic_base;
-
-    set(0xFFFFFFFFFFFFFFFFULL);
-    CHECK(get() == 0xFFFFFFFFFFFFFFFFULL);
-    dump(0);
-}
-
-TEST_CASE("ia32_apic_base_bsp_flag")
-{
-    MockRepository mocks;
-    setup_intrinsics(mocks);
-
-    using namespace msrs::ia32_apic_base;
-
-    bsp_flag::enable();
-    CHECK(bsp_flag::is_enabled());
-    bsp_flag::disable();
-    CHECK(bsp_flag::is_disabled());
-
-    bsp_flag::enable(bsp_flag::mask);
-    CHECK(bsp_flag::is_enabled(bsp_flag::mask));
-    bsp_flag::disable(0x0);
-    CHECK(bsp_flag::is_disabled(0x0));
-}
-
-TEST_CASE("ia32_apic_base_enable_x2apic")
-{
-    MockRepository mocks;
-    setup_intrinsics(mocks);
-
-    using namespace msrs::ia32_apic_base;
-
-    enable_x2apic::enable();
-    CHECK(enable_x2apic::is_enabled());
-    enable_x2apic::disable();
-    CHECK(enable_x2apic::is_disabled());
-
-    enable_x2apic::enable(enable_x2apic::mask);
-    CHECK(enable_x2apic::is_enabled(enable_x2apic::mask));
-    enable_x2apic::disable(0x0);
-    CHECK(enable_x2apic::is_disabled(0x0));
-}
-
-TEST_CASE("ia32_apic_base_apic_global_enable")
-{
-    MockRepository mocks;
-    setup_intrinsics(mocks);
-
-    using namespace msrs::ia32_apic_base;
-
-    apic_global_enable::enable();
-    CHECK(apic_global_enable::is_enabled());
-    apic_global_enable::disable();
-    CHECK(apic_global_enable::is_disabled());
-
-    apic_global_enable::enable(apic_global_enable::mask);
-    CHECK(apic_global_enable::is_enabled(apic_global_enable::mask));
-    apic_global_enable::disable(0x0);
-    CHECK(apic_global_enable::is_disabled(0x0));
-}
-
-TEST_CASE("ia32_apic_base_apic_base")
-{
-    MockRepository mocks;
-    setup_intrinsics(mocks);
-
-    using namespace msrs::ia32_apic_base;
-
-    apic_base::set(0xFFFFFFFFFFFFFFFFULL);
-    CHECK(apic_base::get() == (apic_base::mask >> apic_base::from));
-
-    apic_base::set(apic_base::mask, 0xFFFFFFFFFFFFFFFFULL);
-    CHECK(apic_base::get(apic_base::mask) == (apic_base::mask >> apic_base::from));
-}
-
 TEST_CASE("ia32_mperf")
 {
     MockRepository mocks;
